Split PersonDetailsViewController in
PersonDetailsViewController and PersonOverviewViewController
