+++ src/main/java/org/synyx/urlaubsverwaltung/person/overview/PersonOverviewViewController.java
@@ -1,12 +1,10 @@
-package org.synyx.urlaubsverwaltung.person.web;
+package org.synyx.urlaubsverwaltung.person.overview;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.synyx.urlaubsverwaltung.account.Account;
@@ -18,12 +16,7 @@ import org.synyx.urlaubsverwaltung.department.DepartmentService;
 import org.synyx.urlaubsverwaltung.department.web.UnknownDepartmentException;
 import org.synyx.urlaubsverwaltung.person.Person;
 import org.synyx.urlaubsverwaltung.person.PersonService;
-import org.synyx.urlaubsverwaltung.person.UnknownPersonException;
-import org.synyx.urlaubsverwaltung.person.basedata.PersonBasedata;
 import org.synyx.urlaubsverwaltung.person.basedata.PersonBasedataService;
-import org.synyx.urlaubsverwaltung.settings.SettingsService;
-import org.synyx.urlaubsverwaltung.workingtime.WorkingTime;
-import org.synyx.urlaubsverwaltung.workingtime.WorkingTimeService;
 
 import java.time.Clock;
 import java.time.LocalDate;
@@ -35,7 +28,6 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Stream;
 
-import static java.lang.String.format;
 import static java.util.Comparator.comparing;
 import static java.util.stream.Collectors.toCollection;
 import static java.util.stream.Collectors.toList;
@@ -45,7 +37,6 @@ import static org.synyx.urlaubsverwaltung.person.Role.DEPARTMENT_HEAD;
 import static org.synyx.urlaubsverwaltung.person.Role.INACTIVE;
 import static org.synyx.urlaubsverwaltung.person.Role.OFFICE;
 import static org.synyx.urlaubsverwaltung.person.Role.SECOND_STAGE_AUTHORITY;
-import static org.synyx.urlaubsverwaltung.person.web.PersonDetailsBasedataDtoMapper.mapToPersonDetailsBasedataDto;
 import static org.synyx.urlaubsverwaltung.security.SecurityRules.IS_PRIVILEGED_USER;
 import static org.synyx.urlaubsverwaltung.util.DateUtil.isBeforeApril;
 
@@ -54,98 +45,46 @@ import static org.synyx.urlaubsverwaltung.util.DateUtil.isBeforeApril;
  */
 @Controller
 @RequestMapping("/web")
-public class PersonDetailsViewController {
+public class PersonOverviewViewController {
 
     private static final String BEFORE_APRIL_ATTRIBUTE = "beforeApril";
-    private static final String PERSON_ATTRIBUTE = "person";
 
     private final PersonService personService;
     private final AccountService accountService;
     private final VacationDaysService vacationDaysService;
     private final DepartmentService departmentService;
-    private final WorkingTimeService workingTimeService;
-    private final SettingsService settingsService;
     private final PersonBasedataService personBasedataService;
     private final Clock clock;
 
     @Autowired
-    public PersonDetailsViewController(PersonService personService, AccountService accountService,
-                                       VacationDaysService vacationDaysService, DepartmentService departmentService,
-                                       WorkingTimeService workingTimeService, SettingsService settingsService,
-                                       PersonBasedataService personBasedataService, Clock clock) {
+    PersonOverviewViewController(PersonService personService, AccountService accountService,
+                                 VacationDaysService vacationDaysService, DepartmentService departmentService,
+                                 PersonBasedataService personBasedataService, Clock clock) {
         this.personService = personService;
         this.accountService = accountService;
         this.vacationDaysService = vacationDaysService;
         this.departmentService = departmentService;
-        this.workingTimeService = workingTimeService;
-        this.settingsService = settingsService;
         this.personBasedataService = personBasedataService;
         this.clock = clock;
     }
 
-    @GetMapping("/person/{personId}")
-    public String showPersonInformation(@PathVariable("personId") Integer personId,
-                                        @RequestParam(value = "year", required = false) Optional<Integer> requestedYear,
-                                        Model model) throws UnknownPersonException {
-
-        final Person person = personService.getPersonByID(personId).orElseThrow(() -> new UnknownPersonException(personId));
-        final Person signedInUser = personService.getSignedInUser();
-
-        if (!departmentService.isSignedInUserAllowedToAccessPersonData(signedInUser, person)) {
-            throw new AccessDeniedException(format("User '%s' has not the correct permissions to access data of user '%s'", signedInUser.getId(), person.getId()));
-        }
-
-        final Integer year = requestedYear.orElseGet(() -> Year.now(clock).getValue());
-
-        model.addAttribute("year", year);
-        model.addAttribute(PERSON_ATTRIBUTE, person);
-
-        final Optional<PersonBasedata> basedataByPersonId = personBasedataService.getBasedataByPersonId(person.getId());
-        if(basedataByPersonId.isPresent()) {
-
-            final PersonDetailsBasedataDto personDetailsBasedataDto = mapToPersonDetailsBasedataDto(basedataByPersonId.get());
-            model.addAttribute("personBasedata", personDetailsBasedataDto);
-        }
-
-        model.addAttribute("departments", departmentService.getAssignedDepartmentsOfMember(person));
-        model.addAttribute("departmentHeadOfDepartments", departmentService.getManagedDepartmentsOfDepartmentHead(person));
-        model.addAttribute("secondStageAuthorityOfDepartments", departmentService.getManagedDepartmentsOfSecondStageAuthority(person));
-
-        final Optional<WorkingTime> maybeWorkingTime = workingTimeService.getWorkingTime(person, LocalDate.now(clock));
-        model.addAttribute("workingTime", maybeWorkingTime.orElse(null));
-        model.addAttribute("federalState", maybeWorkingTime.map(WorkingTime::getFederalState)
-            .orElseGet(() -> settingsService.getSettings().getWorkingTimeSettings().getFederalState()));
-
-        model.addAttribute("canEditBasedata", signedInUser.hasRole(OFFICE));
-        model.addAttribute("canEditPermissions", signedInUser.hasRole(OFFICE));
-        model.addAttribute("canEditDepartments", signedInUser.hasRole(OFFICE));
-        model.addAttribute("canEditAccounts", signedInUser.hasRole(OFFICE));
-        model.addAttribute("canEditWorkingtime", signedInUser.hasRole(OFFICE));
-
-        final Optional<Account> maybeAccount = accountService.getHolidaysAccount(year, person);
-        maybeAccount.ifPresent(account -> model.addAttribute("account", account));
-
-        return "person/person_detail";
-    }
-
     @PreAuthorize(IS_PRIVILEGED_USER)
     @GetMapping("/person")
-    public String showPerson() {
+    public String getPersonOverview() {
         return "redirect:/web/person?active=true";
     }
 
     @PreAuthorize(IS_PRIVILEGED_USER)
     @GetMapping(value = "/person", params = "active")
-    public String showPerson(@RequestParam(value = "active") boolean active,
-                             @RequestParam(value = "department", required = false) Optional<Integer> requestedDepartmentId,
-                             @RequestParam(value = "year", required = false) Optional<Integer> requestedYear,
-                             Model model) throws UnknownDepartmentException {
+    public String getPersonOverview(@RequestParam(value = "active") boolean active,
+                                    @RequestParam(value = "department", required = false) Optional<Integer> requestedDepartmentId,
+                                    @RequestParam(value = "year", required = false) Optional<Integer> requestedYear,
+                                    Model model) throws UnknownDepartmentException {
 
         final Integer year = requestedYear.orElseGet(() -> Year.now(clock).getValue());
 
         final Person signedInUser = personService.getSignedInUser();
         final List<Person> persons = active ? getRelevantActivePersons(signedInUser) : getRelevantInactivePersons(signedInUser);
-
 
         if (requestedDepartmentId.isPresent()) {
             final Integer departmentId = requestedDepartmentId.get();
@@ -278,7 +217,6 @@ public class PersonDetailsViewController {
                 .ifPresent(personBasedata -> personDtoBuilder.personnelNumber(personBasedata.getPersonnelNumber()));
 
             final PersonDto personDto = personDtoBuilder.build();
-
             personDtos.add(personDto);
         }
